extracted loading bar ionto separate component
