+++ frontend/src/components/uploadAudioBox.svelte
@@ -1,102 +1,46 @@
 <!--
 
-    Upload Audio Box Component
+  Upload Audio Box Component
 
-    Authors: Parul Garg (pgar0011)
-    Editied by: Benjamin Cherian, Zihao Wang, Angelina Leung
-    Last Modified: 13/05/24
+  consists of parts
+  1. the box itself = blue on initial, green on complete
+  2. dropzone = ideally clear
+  3. description = a text box outline either instruction or completion, 
+    includes an icon, a bolded first line, following lines
+
+  Authors: Parul Garg (pgar0011)
+  Editied by: Benjamin Cherian, Zihao Wang, Angelina Leung
+  Last Modified: 28/07/24
 
 -->
 <!-- JavaScript -->
 <script>
-  import micIcon from "../assets/mic-icon.png";
-  import uploadIcon from "../assets/upload-icon.png";
-  import radioIcon from "../assets/radio-icon.png";
-  import fileIcon from "../assets/file-icon.png";
+  import micIcon from "../assets/mic-icon.png"
   import Dropzone from "svelte-file-dropzone";
+  import LoadingBar from "./loadingBar.svelte";
   import { simple_pathway } from "../api-functions/simple_pathway";
   import { api_status } from "../stores/simple-pathway-store";
   import { goto } from "$app/navigation";
   export let simple = false;
 
+  let loadingBarComponent;
+
+  /**
+   * Subscribes to updates from the `api_status` observable
+   * This function also updates the `loadingBar` with the current status value.
+   */
   const unsubscribe = api_status.subscribe((value) => {
     console.log(value.status); // Log the status value
-    updateLoadingBar(value.status);
+    console.log(typeof updateLoadingBar); // Should log 'function'
+    if (loadingBarComponent) {
+      loadingBarComponent.updateLoadingBar(value.status);
+    }
   });
-
-  function updateLoadingBar(api_status) {
-    if (api_status == "Transcribe") {
-      updateUploadBoxContents("Transcribing audio", false);
-      currentProgress = 0.4;
-      makeProgression();
-
-      setTimeout(() => {
-        currentProgress = 0.55;
-        makeProgression();
-      }, 8000);
-
-    } else if (api_status == "Summary") {
-      updateUploadBoxContents("Generating summary", true);
-      currentProgress = 0.7;
-      makeProgression();
-
-    } else if (api_status == "Email") {
-      updateUploadBoxContents("Sending email", true);
-      currentProgress = 0.9;
-      makeProgression();
-
-    } else if (api_status == "Complete") {
-      currentProgress = 0.99;
-      makeProgression();
-      setTimeout(() => {
-        goto("/sent");
-      }, 2500);
-      
-    }
-  }
-
-  // ------------------------------------------ Progress Bar
-  let progressBarWidth = 0;
-  let progressBarProgress;
-  let progressBarDisplay = 0;
-  let loadingBarWidth;
-  let progressBarMax = 1;
-  let currentProgress = 0;
-
-  $: switch (progressBarWidth) {
-    case Math.round(loadingBarWidth * progressBarMax):
-      clearInterval(progressBarProgress);
-  }
-
-  const progression = () => {
-    if (progressBarWidth < currentProgress * loadingBarWidth) {
-      progressBarWidth += 1;
-      progressBarDisplay = Math.round(
-        (progressBarWidth / loadingBarWidth) * 100,
-      );
-    }
-  };
-
-  const makeProgression = () => {
-    loadingBarWidth = document.getElementById("loadingBar").clientWidth;
-    document.getElementById("loadingBar").style.visibility = "visible";
-    document.getElementById("progressNumber").style.visibility = "visible";
-    progressBarProgress = setInterval(progression, 10);
-
-    return true;
-  };
-  const MAX_DURATION_SECONDS = 7200; // 7200 seconds = 120 minutes
-
-  var ICON_DICT = {
-    "Uploading meeting audio": uploadIcon,
-    "Transcribing audio": radioIcon,
-    "Generating summary": fileIcon,
-    "Sending email": fileIcon,
-  };
 
   // ------------------------------------------ File Handling
   let file;
   let errorMessage = "";
+  const MAX_DURATION_SECONDS = 7200; // 7200 seconds = 120 minutes, used to check limit on files
 
   async function handleFilesSelect(e) {
     const { acceptedFiles } = e.detail;
@@ -135,47 +79,6 @@
         "Invalid audio format! Your meeting audio must be in MP3 or WAV format";
     }
   }
-
-  // ------------------------------------------ Box Contents
-  // changes the contents of the audio box
-  function updateUploadBoxContents(newText, isProgression) {
-    var firstLine = document.querySelector(".first-line");
-    // changeClass('.first-line', '.loading-line')
-    firstLine.textContent = newText + "...";
-
-    if (isProgression) {
-      // Change icon source
-      var icon = document.getElementById("icon");
-      icon.src = ICON_DICT[newText];
-    } else {
-      var secondLine = document.querySelector(".second-line");
-      changeClass(".second-line", "loading-line");
-      secondLine.textContent = "";
-
-      // Change icon source
-      var icon = document.querySelector(".large-icon");
-      icon.src = ICON_DICT[newText];
-      changeClass(".large-icon", ".small-icon");
-    }
-
-    // when a file is dragged in show the loading bar and 'Uploading meeting audio'
-    // when file is sent to assemblyai; show 'Transcribing audio'
-    // when file is sent to gemini; show 'Generating Summary'
-    // when file is sent to assemblyai; show 'transcribing audio'
-  }
-
-  function changeClass(elemId, newClassName) {
-    // Get the span element by its ID
-    const spanElement = document.querySelector(elemId);
-
-    // Remove any existing class from the span element
-    if (spanElement != null) {
-      spanElement.className = "";
-
-      // Add the new class to the span element
-      spanElement.classList.add(newClassName);
-    }
-  }
 </script>
 
 <!-- COMPONENT -->
@@ -185,15 +88,9 @@
       <!-- The dropzone is on top of custom-input so the grey is covering the lightblue-->
       <img id="icon" class="large-icon" src={micIcon} alt="Icon" />
       <span class="first-line">Upload meeting audio</span>
-      <span class="second-line"
-        >Must be under 120 minutes. MP3 or WAV formats accepted.</span
-      >
-
-      <div id="loadingBar">
-        <div id="progressBar" style="width: {progressBarWidth}px"></div>
-      </div>
-      <div id="progressNumber"><b>{progressBarDisplay}</b>%</div>
+      <span class="second-line">Must be under 120 minutes. MP3 or WAV formats accepted.</span>
     </Dropzone>
+    <LoadingBar bind:this={loadingBarComponent}/>
   </label>
 </div>
